Atualização do log + otimização de precisão

+++ mandelbrot.c
@@ -2,26 +2,26 @@
 #include <stdlib.h>
 #include "mandelbrot.h"
 
-float min_real = -2.0;
-float max_real = 1.0;
-float min_imag = -1.5;
-float max_imag = 1.5;
+double min_real = -2.0;
+double max_real = 1.0;
+double min_imag = -1.5;
+double max_imag = 1.5;
 
 int calcular_pixel_mandelbrot(int x, int largura, int y, int altura, int max_iter){
 
-    float real = (max_real - min_real)/largura;
-    float imag = (max_imag - min_imag)/altura;
+    double real = (max_real - min_real)/largura;
+    double imag = (max_imag - min_imag)/altura;
 
-    float c_real = min_real + (x * real);
-    float c_imag = min_imag + (y * imag);
+    double c_real = min_real + (x * real);
+    double c_imag = min_imag + (y * imag);
 
-    float a = 0.0;
-    float b = 0.0;
+    double a = 0.0;
+    double b = 0.0;
     int iteracoes = 0;
 
     while ((a * a + b * b <= 4.0) && (iteracoes < max_iter)){
-        float a_novo = (a * a) - (b * b) + c_real;
-        float b_novo = 2*a*b + c_imag;
+        double a_novo = (a * a) - (b * b) + c_real;
+        double b_novo = 2*a*b + c_imag;
 
         a = a_novo;
         b = b_novo;
